classifieds::parse: Split module up further
