+++ src/classifieds/parse/mod.rs
@@ -1,0 +1,5 @@
+mod item;
+mod user;
+
+pub use item::ClassifiedsDetails;
+pub use user::ClassifiedsUser;
